+++ CLAUDE.md
@@ -19,23 +19,36 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 - `just format` - Format code (requires .ci/format.jl)
 - `just changelog` - Generate changelog (requires .ci/changelog.jl)
 
+## Code Organization
+
+### File Structure
+REPLicant follows a modular design with source files organized by functionality:
+
+- **Main module** (`src/REPLicant.jl`) - Module definition, imports, and includes only
+- **Server** (`src/server.jl`) - Socket server implementation and client handling
+- **Evaluation** (`src/evaluation.jl`) - Code evaluation and command dispatch
+- **Metadata commands** (`src/metadata_commands.jl`) - All #meta command implementations
+- **Metadata helpers** (`src/metadata_helpers.jl`) - Helper functions for metadata inspection
+- **Formatting** (`src/formatting.jl`) - Output formatting utilities
+- **Justfile support** (`src/justfile_support.jl`) - Justfile integration
+- **Filesystem** (`src/filesystem.jl`) - File system utilities
+
+### Code Style Guidelines
+- **Keep files focused and manageable** - Each file should have a single, clear purpose
+- **Target 200-500 lines per file** - Split larger files into logical components
+- **Use clear section headers** - Organize code with commented sections (e.g., `# Section name`)
+- **Group related functions** - Keep functions that work together in the same file
+- **Minimize cross-file dependencies** - Reduce coupling between modules
+
 ## Architecture
 
 REPLicant is a socket-based Julia code evaluation server that allows remote execution of Julia code:
 
-1. **Server Component** (`src/REPLicant.jl`)
-   - Creates a socket server that listens on a dynamic port (starting from 8000)
-   - Writes the port number to `REPLICANT_PORT` file in the project directory
-   - Accepts code strings from clients, evaluates them, and returns results
-   - Integrates with Revise.jl for automatic code reloading via extension
+1. **Server Component** - Creates a socket server that listens on a dynamic port (starting from 8000), writes the port number to `REPLICANT_PORT` file, accepts code strings from clients, evaluates them, and returns results
 
-2. **Revise Extension** (`ext/REPLicantReviseExt.jl`)
-   - Provides automatic code reloading when Revise.jl is loaded
-   - Checks revision queue before executing code and runs `revise()` if needed
+2. **Revise Extension** (`ext/REPLicantReviseExt.jl`) - Provides automatic code reloading when Revise.jl is loaded, checks revision queue before executing code and runs `revise()` if needed
 
-3. **Client Interface** (`justfile`)
-   - Uses netcat (`nc`) to send code to the server
-   - Reads port from `REPLICANT_PORT` file to connect to the correct server instance
+3. **Client Interface** (`justfile`) - Uses netcat (`nc`) to send code to the server, reads port from `REPLICANT_PORT` file to connect to the correct server instance
 
 The server captures both stdout and return values using IOCapture, handles errors gracefully, and formats output similar to the Julia REPL.
