+++ Project.toml
@@ -1,7 +1,7 @@
 name = "REPLicant"
 uuid = "8acbef56-6722-412a-9f59-5674813d37bb"
-authors = ["MichaelHatherly <michaelhatherly@gmail.com>"]
 version = "1.1.0"
+authors = ["MichaelHatherly <michaelhatherly@gmail.com>"]
 
 [deps]
 IOCapture = "b5f81e59-6552-4d32-b1f0-c071b021bf89"
